refactor: move routing fns into wss.ts; replace reserveTaskForWorker with assignIdleWorkers
Per reviewer feedback:
- Delete event-routing.ts; all routing free functions (routeEvent, forwardEvent,
  routePrEvent, etc.) and RoutingDeps now live in wss.ts alongside the hello
  handlers, keeping the module cohesive
- Replace TaskManager.reserveTaskForWorker with assignIdleWorkers(): mutex now
  wraps the entire idle-worker loop; emits 'assigned' / 'assign_failed' events
  so ForemanWss can call sendMsg without a callback crossing the model/controller
  boundary
- RoutingDeps no longer carries assignIdleWorkers; routing functions call
  deps.taskManager.assignIdleWorkers() directly
- Update all tests to import from wss.js and test the new event-based API
