+++ CLAUDE.md
@@ -10,8 +10,8 @@ Two independent processes: **foreman** (server) and **worker** (agent). They com
 
 MVC structure. Models own state; controllers handle external inputs.
 
-- **Models** (`models/`) — `Task` is the active-record for tasks. `WebhookEvent` is the active-record for incoming GitHub webhook events (`webhook_events` table); `WebhookEvent.fromIncoming()` builds an in-memory instance and `toWorkerPayload()` maps `eventName` → `name` for the wire protocol. `ForemanMessage` is the active-record for foreman↔worker messages (`foreman_messages` table); `ForemanMessage.buildSummary()` is the single source of truth for log entry summaries. `queryActivityLog()` (in `activity-log.ts`) merges both tables by timestamp. `TaskManager` owns ephemeral in-memory state only (event queues, branch mappings, blocker state) and exposes `reserveTaskForWorker(workerId)` — an atomic find-and-assign with an internal mutex that prevents double-assignment. `Worker` is the active-record for connected workers (module-level registry, static finders, `Worker._reset()` for test isolation).
-- **Controllers** (`controllers/`) — `http-server.ts` handles webhooks + REST + SPA. `wss.ts` is the protocol layer: handles both directions of the worker WebSocket channel and webhook routing (by delegating to free functions in `event-routing.ts`). `event-routing.ts` contains the free functions (`routeEvent`, `forwardEvent`, `routePrEvent`, etc.) and the `RoutingDeps` interface — imported by both `wss.ts` and tests.
+- **Models** (`models/`) — `Task` is the active-record for tasks. `WebhookEvent` is the active-record for incoming GitHub webhook events (`webhook_events` table); `WebhookEvent.fromIncoming()` builds an in-memory instance and `toWorkerPayload()` maps `eventName` → `name` for the wire protocol. `ForemanMessage` is the active-record for foreman↔worker messages (`foreman_messages` table); `ForemanMessage.buildSummary()` is the single source of truth for log entry summaries. `queryActivityLog()` (in `activity-log.ts`) merges both tables by timestamp. `TaskManager` owns ephemeral in-memory state only (event queues, branch mappings, blocker state) and exposes `assignIdleWorkers()` — assigns all pending tasks to all idle workers with an internal mutex preventing double-assignment; emits `'assigned'` and `'assign_failed'` events. `Worker` is the active-record for connected workers (module-level registry, static finders, `Worker._reset()` for test isolation).
+- **Controllers** (`controllers/`) — `http-server.ts` handles webhooks + REST + SPA. `wss.ts` is the protocol layer: handles both directions of the worker WebSocket channel and webhook routing. It exports the free functions (`routeEvent`, `forwardEvent`, `routePrEvent`, etc.), the `RoutingDeps` interface, and the hello handlers (`handleBusyHello`, `handleIdleHello`) — all imported by tests.
 - **Infrastructure** — `db-client.ts` wires the shared Supabase client and exports `DbRow<T>`, a helper type that makes `id` optional for in-memory (unsaved) model instances. `admin-ws.ts` broadcasts to the admin dashboard. `github.ts` wraps GitHub API calls.
 
 ### Agent/Worker (`src/agent/`)
@@ -86,5 +86,5 @@ See **`docs/type-system.md`** for the full design. In brief: one server model cl
 - In unit tests, use `setupInMemoryTasks()` from `tests/helpers/task.ts` instead of `initDb()` — it mocks the `Task` layer with an in-memory map. `WebhookEvent.log()` and `ForemanMessage.log()` return `Promise<void>`; production callers use `void` to discard the promise; tests can `await` them for deterministic setup. `tests/setup.ts` (a vitest `setupFiles` entry) always initialises the DB with the in-memory shim from `tests/helpers/memory-db.ts` so that these calls no-op cleanly without guards in production code; DB tests override this with `initDb(createTestSupabase())` at module level.
 - All wire types live in `shared/wire.ts` and are imported as `import * as Wire from "../../shared/wire.js"` (adjust relative path as needed). `src/wire.ts` re-exports from `shared/wire.ts` for backward compatibility. The DB model class for foreman messages is `ForemanMessage` (in `src/foreman/models/foreman-message.ts`), distinct from the wire type `Wire.ForemanMessage`.
 - In agent tests that need slash commands, create a `CommandRegistry` instance directly (`new CommandRegistry()`) and populate it via `registerTestCommands()` from `tests/helpers.ts` (which returns the registry). Pass the registry to functions like `dispatchInput`, `parseSlashCommand`, etc.; call `registry.listCommandNames()` and `registry.listCommands()` directly on the registry (with injectable `listDir`/`readFile` args for testing). When calling `registerWorkspaceCommands` or `registerWorkerCommands` in tests, pass a pre-scoped registry (e.g. `registry.scoped("workspace")`) — scoping is always the caller's responsibility.
-- In foreman tests that use `new ForemanWss(...)`, call `Worker._reset()` in `beforeEach` to clear the module-level worker registry between tests. `ForemanWss` takes a single options object `{ config, taskManager, server, adminWss? }` — `Worker` is imported directly by `wss.ts` and `event-routing.ts` (no registry injection). To unit-test the busy/idle reconnection logic without a WebSocket server, import `handleBusyHello` / `handleIdleHello` from `wss.ts` and call them directly with mocked deps. To unit-test event routing, import the free functions (`routeEvent`, `forwardEvent`, etc.) from `event-routing.ts` and construct a `RoutingDeps` object directly.
+- In foreman tests that use `new ForemanWss(...)`, call `Worker._reset()` in `beforeEach` to clear the module-level worker registry between tests. `ForemanWss` takes a single options object `{ config, taskManager, server, adminWss? }` — `Worker` is imported directly by `wss.ts` (no registry injection). To unit-test the busy/idle reconnection logic without a WebSocket server, import `handleBusyHello` / `handleIdleHello` from `wss.ts` and call them directly with mocked deps. To unit-test event routing, import the free functions (`routeEvent`, `forwardEvent`, etc.) from `wss.ts` and construct a `RoutingDeps` object directly (note: `RoutingDeps` does not include `assignIdleWorkers` — routing functions call `deps.taskManager.assignIdleWorkers()` directly, so mock that on the taskManager object).
 - In browser tests, the server is shared across all tests — use unique issue numbers per test rather than resetting server state.
