+++ src/foreman/controllers/wss.ts
@@ -11,7 +11,7 @@ import type { BrunelConfig } from "../../config.js";
 import type { TaskManager } from "../models/task-manager.js";
 import { Task } from "../models/task.js";
 import { Worker } from "../models/worker.js";
-import { routeEvent } from "./event-routing.js";
+import { fetchIssueStates } from "../github.js";
 
 type R = Record<string, unknown>;
 
@@ -21,6 +21,297 @@ function debounce(fn: () => void | Promise<void>, delayMs: number): () => void {
     if (timer !== null) clearTimeout(timer);
     timer = setTimeout(() => { timer = null; void fn(); }, delayMs);
   };
+}
+
+// ── Routing helpers ───────────────────────────────────────────────────────────
+
+function strProp(obj: unknown, key: string): string | null {
+  if (typeof obj !== "object" || obj === null) return null;
+  const val = (obj as Record<string, unknown>)[key];
+  return typeof val === "string" ? val : null;
+}
+
+function numProp(obj: unknown, key: string): number | null {
+  if (typeof obj !== "object" || obj === null) return null;
+  const val = (obj as Record<string, unknown>)[key];
+  return typeof val === "number" ? val : null;
+}
+
+function extractLinkedIssueNumber(body: string): number | null {
+  const match = /(?:closes|fixes|resolves)\s+#(\d+)/i.exec(body);
+  return match ? parseInt(match[1], 10) : null;
+}
+
+// ── Event routing (exported for unit testing) ─────────────────────────────────
+
+export interface RouteResult { taskId: string | null; workerId: string | null; }
+
+export interface RoutingDeps {
+  taskManager: TaskManager;
+  repo: string;
+  token: string;
+  githubApiUrl?: string;
+  taskLabel: string;
+  sendMsg(workerId: string, msg: Wire.ForemanMessage, logTaskId?: string): void;
+  flog(msg: string): void;
+}
+
+export function forwardEvent(task: Task, evt: WebhookEvent, ref: string, deps: Pick<RoutingDeps, "sendMsg" | "flog" | "taskManager">): void {
+  const { sendMsg, flog, taskManager } = deps;
+  if (task.workerId) {
+    const worker = Worker.get(task.workerId);
+    if (worker && worker.currentTaskId !== task.taskId) {
+      flog(`[task ${ref}] ${evt.eventName} dropped — worker ${shortWorkerId(task.workerId)} is now on a different task`);
+      return;
+    }
+    if (worker?.status === "disconnected") {
+      taskManager.queueEvent(task.taskId, evt);
+      flog(`[task ${ref}] ${evt.eventName} queued (worker ${shortWorkerId(task.workerId)} disconnected)`);
+    } else if (worker) {
+      const evtMsg: Wire.ForemanMessage = { type: "event_notification", taskId: task.taskId, event: evt.toWorkerPayload() };
+      sendMsg(task.workerId, evtMsg);
+      flog(`[worker ${shortWorkerId(task.workerId)}] → event_notification ${ref} ${evt.eventName}`);
+    } else {
+      flog(`[task ${ref}] ${evt.eventName} DROPPED — worker ${shortWorkerId(task.workerId)} not in registry (disconnected?)`);
+    }
+  } else if (task.status === "pending" || task.status === "blocked") {
+    taskManager.queueEvent(task.taskId, evt);
+    flog(`[task ${ref}] ${evt.eventName} queued (no worker assigned)`);
+  }
+}
+
+function startDepsLoad(issueNumber: number, body: string, deps: RoutingDeps): void {
+  const { taskManager, repo, token, githubApiUrl, flog } = deps;
+  Task.fetchBlockers(issueNumber, body, { repo, token, apiUrl: githubApiUrl })
+    .then(async (blockers) => {
+      taskManager.setBlockers(issueNumber, blockers);
+      if (blockers.length > 0) {
+        const states = await fetchIssueStates(blockers, { repo, token });
+        for (const [num, state] of states) {
+          taskManager.setIssueOpenState(num, state === "open");
+        }
+      }
+      taskManager.markBlockersLoaded(issueNumber);
+      await taskManager.assignIdleWorkers();
+    })
+    .catch((err) => flog(`ERROR fetching deps for #${issueNumber}: ${fmtError(err)}`));
+}
+
+export async function routePrEvent(
+  p: Record<string, unknown>,
+  evt: WebhookEvent,
+  deps: RoutingDeps,
+): Promise<RouteResult> {
+  function result(task: { taskId: string; workerId: string | null } | null | undefined): RouteResult {
+    return { taskId: task?.taskId ?? null, workerId: task?.workerId ?? null };
+  }
+
+  const { taskManager, flog } = deps;
+  const pr = p.pull_request as Record<string, unknown> | undefined;
+  const prNumber = numProp(pr, "number");
+  if (prNumber === null) return result(null);
+
+  if (p.action === "synchronize") return result(await Task.getByPr(prNumber));
+
+  if (p.action === "opened" && pr) {
+    const linkedIssue = extractLinkedIssueNumber(String(pr.body ?? ""));
+    if (linkedIssue !== null) {
+      const linkedTask = await Task.getByIssue(linkedIssue);
+      if (linkedTask) {
+        const branch = strProp(pr.head, "ref");
+        if (branch) taskManager.registerBranch(branch, linkedTask.taskId);
+        await linkedTask.registerPr(prNumber, branch ?? null).catch((err: unknown) =>
+          flog(`ERROR Failed to register PR for task #${linkedTask.taskId}: ${fmtError(err)}`)
+        );
+        flog(`[task #${linkedIssue}] PR #${prNumber} registered`);
+      }
+    }
+  }
+
+  if (p.action === "closed" && pr && !pr.merged) {
+    const task = await Task.getByPr(prNumber);
+    if (task) {
+      flog(`[task #${task.issueNumber}] PR #${prNumber} unregistered (closed without merging)`);
+      await task.unregisterPr().catch((err: unknown) =>
+        flog(`ERROR Failed to unregister PR #${prNumber}: ${fmtError(err)}`)
+      );
+      forwardEvent(task, evt, `PR #${prNumber}`, deps);
+      return result(task);
+    }
+    return result(null);
+  }
+
+  if (p.action === "closed" && pr && pr.merged) {
+    const task = await Task.getByPr(prNumber);
+    if (task) {
+      flog(`[task #${task.issueNumber}] PR #${prNumber} merged`);
+      await task.mergePr().catch((err: unknown) =>
+        flog(`ERROR Failed to record PR #${prNumber} merge: ${fmtError(err)}`)
+      );
+      forwardEvent(task, evt, `PR #${prNumber}`, deps);
+      return result(task);
+    }
+    return result(null);
+  }
+
+  const task = await Task.getByPr(prNumber);
+  if (task) forwardEvent(task, evt, `PR #${prNumber}`, deps);
+  return result(task);
+}
+
+export async function routePrReviewEvent(
+  p: Record<string, unknown>,
+  evt: WebhookEvent,
+  deps: RoutingDeps,
+): Promise<RouteResult> {
+  function result(task: { taskId: string; workerId: string | null } | null | undefined): RouteResult {
+    return { taskId: task?.taskId ?? null, workerId: task?.workerId ?? null };
+  }
+
+  const pr = p.pull_request as Record<string, unknown> | undefined;
+  const prNumber = numProp(pr, "number");
+  if (prNumber === null) return result(null);
+  const task = await Task.getByPr(prNumber);
+  if (task) forwardEvent(task, evt, `PR #${prNumber}`, deps);
+  return result(task);
+}
+
+export async function routeCheckEvent(
+  p: Record<string, unknown>,
+  evt: WebhookEvent,
+  name: string,
+  deps: RoutingDeps,
+): Promise<RouteResult> {
+  function result(task: { taskId: string; workerId: string | null } | null | undefined): RouteResult {
+    return { taskId: task?.taskId ?? null, workerId: task?.workerId ?? null };
+  }
+
+  const { taskManager } = deps;
+  const inner = (name === "check_run" ? p.check_run : p.check_suite) as Record<string, unknown> | undefined;
+  const prs = inner?.pull_requests as Array<{ number: number }> | undefined;
+
+  if (prs && prs.length > 0) {
+    const task = await Task.getByPr(prs[0].number);
+    if (task) { forwardEvent(task, evt, `PR #${prs[0].number}`, deps); return result(task); }
+  }
+
+  const headBranch = name === "check_run"
+    ? strProp(inner?.check_suite, "head_branch") ?? ""
+    : strProp(inner, "head_branch") ?? "";
+  if (headBranch) {
+    const task = await taskManager.getTaskForBranch(headBranch);
+    if (task) { forwardEvent(task, evt, `branch ${headBranch}`, deps); return result(task); }
+  }
+  return result(null);
+}
+
+export async function routeIssueEvent(
+  p: Record<string, unknown>,
+  evt: WebhookEvent,
+  issue: Record<string, unknown>,
+  issueNumber: number,
+  deps: RoutingDeps,
+): Promise<RouteResult> {
+  function result(task: { taskId: string; workerId: string | null } | null | undefined): RouteResult {
+    return { taskId: task?.taskId ?? null, workerId: task?.workerId ?? null };
+  }
+
+  const { taskManager, flog, repo, taskLabel } = deps;
+
+  let task = await Task.getByIssue(issueNumber);
+  if (!task) task = await Task.getByPr(issueNumber);
+
+  if (!task) {
+    const action = p.action as string | undefined;
+    const labeledNow =
+      action === "labeled" &&
+      (p.label as Record<string, unknown> | undefined)?.name === taskLabel;
+    const openedWithLabel =
+      action === "opened" &&
+      (issue.labels as Array<{ name: string }> | undefined)?.some((l) => l.name === taskLabel);
+
+    if (labeledNow || openedWithLabel) {
+      const issueState = String(issue.state ?? "open");
+      if (issueState === "closed") {
+        flog(`[task #${issueNumber}] issues/${action}: ignoring — issue is closed (title: ${JSON.stringify(String(issue.title ?? ""))})`);
+        return result(null);
+      }
+
+      const labels = (issue.labels as Array<{ name: string }> | undefined)?.map((l) => l.name) ?? [];
+      await taskManager.enqueueIssue(String(issueNumber), issueNumber, repo, String(issue.title ?? ""), String(issue.body ?? ""), labels)
+        .catch((err: unknown) => flog(`ERROR Failed to persist task #${issueNumber}: ${fmtError(err)}`));
+
+      startDepsLoad(issueNumber, String(issue.body ?? ""), deps);
+      await taskManager.assignIdleWorkers();
+      flog(`[task #${issueNumber}] enqueued via issues/${action}`);
+      return { taskId: String(issueNumber), workerId: null };
+    }
+  }
+
+  const action = p.action as string | undefined;
+
+  if (
+    action === "unlabeled" &&
+    (p.label as Record<string, unknown> | undefined)?.name === taskLabel
+  ) {
+    await taskManager.dequeueIssue(issueNumber)
+      .catch((err: unknown) => flog(`ERROR Failed to dequeue task #${issueNumber}: ${fmtError(err)}`));
+    flog(`[task #${issueNumber}] dequeued (label removed)`);
+    await taskManager.assignIdleWorkers();
+    return result(task);
+  }
+
+  if (action === "closed") {
+    await taskManager.closeIssue(issueNumber).catch((err: unknown) =>
+      flog(`ERROR Failed to close issue #${issueNumber}: ${fmtError(err)}`)
+    );
+    await taskManager.assignIdleWorkers();
+    return result(task);
+  }
+
+  if (action === "reopened") {
+    await taskManager.reopenIssue(issueNumber).catch((err: unknown) =>
+      flog(`ERROR Failed to reopen issue #${issueNumber}: ${fmtError(err)}`)
+    );
+    await taskManager.assignIdleWorkers();
+    return result(task);
+  }
+
+  if (action === "edited") {
+    const changes = p.changes as Record<string, unknown> | undefined;
+    if (changes?.body) {
+      const trackedTask = await Task.getByIssue(issueNumber);
+      if (trackedTask) {
+        const newBody = String(issue.body ?? "");
+        taskManager.resetBlockers(issueNumber);
+        startDepsLoad(issueNumber, newBody, deps);
+      }
+    }
+  }
+
+  if (!task) return result(null);
+  forwardEvent(task, evt, `#${issueNumber}`, deps);
+  return result(task);
+}
+
+export async function routeEvent(name: string, p: Record<string, unknown>, evt: WebhookEvent, deps: RoutingDeps): Promise<RouteResult> {
+  if (name === "pull_request") {
+    return routePrEvent(p, evt, deps);
+  }
+
+  if (name === "pull_request_review" || name === "pull_request_review_comment") {
+    return routePrReviewEvent(p, evt, deps);
+  }
+
+  if (name === "check_run" || name === "check_suite") {
+    return routeCheckEvent(p, evt, name, deps);
+  }
+
+  const issue = p.issue as Record<string, unknown> | undefined;
+  const issueNumber = numProp(issue, "number");
+  if (issueNumber === null) return { taskId: null, workerId: null };
+
+  return routeIssueEvent(p, evt, issue!, issueNumber, deps);
 }
 
 // ── Busy/idle hello handlers (exported for unit testing) ─────────────────────
@@ -70,11 +361,9 @@ async function reclaimWorker(
   const { ws, sendMsg } = deps;
   const w = Worker.register(workerId, ws);
   w.assign(task.taskId);
-  // Only call assign if task is not already complete (to preserve task status)
   if (task.status !== "complete") {
     await task.assign(workerId);
   }
-  // For complete tasks, the task stays complete while worker finishes cleanup/finalization work
   sendMsg(workerId, { type: "hello_ack", workerId, status: "busy" }, task.taskId);
   flushQueuedEvents(workerId, task.taskId, task.issueNumber, deps);
 }
@@ -217,27 +506,8 @@ export class ForemanWss {
     taskManager.on("changed", debouncedBroadcast);
     Worker.events.on("changed", debouncedBroadcast);
 
-    // Sequential (not concurrent) to prevent double-assignment: TaskManager.reserveTaskForWorker
-    // uses an internal mutex so each DB write completes before the next call queries nextPending().
-    // (Issue #563, #577)
-    const assignIdleWorkers = async (): Promise<void> => {
-      for (const w of Worker.getIdle()) {
-        await tryAssignWork(w.workerId).catch(err => flog(`ERROR tryAssignWork: ${fmtError(err)}`));
-      }
-    };
-
-    const tryAssignWork = async (workerId: string): Promise<void> => {
-      let result: { task: Task; queued: WebhookEvent[] } | null;
-      try {
-        result = await taskManager.reserveTaskForWorker(workerId);
-      } catch (err) {
-        flog(`ERROR Failed to persist assignment: ${fmtError(err)}`);
-        Worker.get(workerId)?.release();
-        log(workerId, "→ idle (DB write failed)");
-        return;
-      }
-      if (!result) return;
-      const { task, queued } = result;
+    // Listen for task assignment outcomes from TaskManager.
+    taskManager.on("assigned", ({ task, queued, workerId: wid }: { task: Task; queued: WebhookEvent[]; workerId: string }) => {
       const assignMsg: Wire.ForemanMessage = {
         type: "task_assigned",
         taskId: task.taskId,
@@ -249,16 +519,21 @@ export class ForemanWss {
           repoUrl: task.repoUrl,
         },
       };
-      sendMsg(workerId, assignMsg);
-      log(workerId, `→ task_assigned #${task.issueNumber} "${task.title}"`);
+      sendMsg(wid, assignMsg);
+      log(wid, `→ task_assigned #${task.issueNumber} "${task.title}"`);
       for (const evt of queued) {
         const evtMsg: Wire.ForemanMessage = { type: "event_notification", taskId: task.taskId, event: evt.toWorkerPayload() };
-        sendMsg(workerId, evtMsg);
-        log(workerId, `→ event_notification #${task.issueNumber} ${evt.eventName} (queued)`);
+        sendMsg(wid, evtMsg);
+        log(wid, `→ event_notification #${task.issueNumber} ${evt.eventName} (queued)`);
       }
-    };
+    });
 
-    const routingDeps = { taskManager, repo, token, githubApiUrl, taskLabel, sendMsg, flog, assignIdleWorkers };
+    taskManager.on("assign_failed", ({ workerId: wid, err }: { workerId: string; err: unknown }) => {
+      flog(`ERROR Failed to persist assignment: ${fmtError(err)}`);
+      log(wid, "→ idle (DB write failed)");
+    });
+
+    const routingDeps: RoutingDeps = { taskManager, repo, token, githubApiUrl, taskLabel, sendMsg, flog };
 
     this.routeEvent = async (id: string, name: string, payload: unknown) => {
       const p = payload as Record<string, unknown>;
@@ -292,7 +567,7 @@ export class ForemanWss {
       });
     };
 
-    this.reconcile = assignIdleWorkers;
+    this.reconcile = () => taskManager.assignIdleWorkers();
 
     const wss = new WebSocketServer({ noServer: true });
     this.wss = wss;
@@ -373,7 +648,7 @@ export class ForemanWss {
           else if (msg.type === "task_complete") await handleTaskComplete(msg);
           else if (msg.type === "worker_goodbye") await handleWorkerGoodbye(msg);
           else { flog(`[worker ${workerId}] unknown message type: ${(msg as R).type}`); return; }
-          await assignIdleWorkers();
+          await taskManager.assignIdleWorkers();
         })().catch(err => flog(`ERROR handling worker message: ${fmtError(err)}`));
       });
